pyproject.toml
[tool.poetry]
name = "pyguis"
version = "0.0.0"
description = "Guis for sql dbs, with support for reading images from http."
authors = ["manyids2 <arun.mukundan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
tqdm = "^4.57.0"
python = "^3.10.0"
numpy = "^1.24.1"
pytest = "^7.2.0"
pyyaml = "^5.0"
black = "^22.6.0"
rich = "^13.0.1"
pandas = "^1.5.1"
dearpygui = "^1.8.0"
pillow = "^9.4.0"
sphinx = "^5.0.0"
sphinxcontrib-bibtex = "^2.5.0"
sphinx-rtd-theme = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
